move library and syncer to git package
