+++ service/git/syncer_test.go
@@ -1,9 +1,10 @@
-package lookout
+package git
 
 import (
 	"context"
 	"testing"
 
+	"github.com/src-d/lookout"
 	"github.com/stretchr/testify/require"
 	"gopkg.in/sourcegraph/go-vcsurl.v1"
 	"gopkg.in/src-d/go-billy.v4/memfs"
@@ -15,8 +16,8 @@ func TestLibrary_Sync(t *testing.T) {
 	syncer := NewSyncer(library)
 
 	url, _ := vcsurl.Parse("http://github.com/src-d/lookout")
-	err := syncer.Sync(context.TODO(), &CommitRevision{
-		Head: ReferencePointer{
+	err := syncer.Sync(context.TODO(), &lookout.CommitRevision{
+		Head: lookout.ReferencePointer{
 			InternalRepositoryURL: url.CloneURL,
 			ReferenceName:         "refs/pull/1/head",
 			Hash:                  "80a9810a027672a098b07efda3dc305409c9329d",
